# src/controllers/auth.js
import {
  registerUser,
  loginUser,
  refreshSession,
  logoutUser,
  requestResetPassword,
  updatePassword,
} from '../sevices/auth.js';
import { saveRefreshCookies } from '../utils/session.js';

export const registerController = async (req, res) => {
  const newUser = await registerUser(req.body);

  res.status(201).json({
    status: 201,
    message: 'Successfully registered a user!',
    data: newUser,
  });
};

export const loginController = async (req, res) => {
  const session = await loginUser(req.body);

  saveRefreshCookies(res, session);

  res.status(200).json({
    status: 200,
    message: 'Successfully logged an user!',
    data: {
      accessToken: session.accessToken,
    },
  });
};

export const refreshController = async (req, res) => {
  const session = await refreshSession(
    req.cookies.sessionId,
    req.cookies.refreshToken,
  );

  saveRefreshCookies(res, session);

  res.status(200).json({
    status: 200,
    message: 'Successfully refreshed a session!',
    data: {
      accessToken: session.accessToken,
    },
  });
};

export const logoutController = async (req, res) => {
  await logoutUser(req.cookies.sessionId);

  res.clearCookie('sessionId');
  res.clearCookie('refreshToken');

  res.status(204).end();
};

export const requestResetPasswordController = async (req, res) => {
  await requestResetPassword(req.body.email);

  res.status(200).json({
    status: 200,
    message: 'Reset password email has been successfully sent.',
    data: {},
  });
};

export const resetPasswordController = async (req, res) => {
  await updatePassword(req.body.password, req.body.token);

  res.status(200).json({
    status: 200,
    message: 'Password has been successfully reset.',
    data: {},
  });
};